Stop ignoring unit tests just because the Mono compiler built them. Peverify checks are still commented out.
With this change, as of Mono 4.2.1, only 9 tests still fail.

The real issue is whether the Mono _runtime_ is running the tests; .NET could run tests compiled on Mono, and Mono could run tests compiled on .NET. If truly necessary, it's possible to ignore tests at runtime by throwing the IgnoreException.

Otherwise, the tests SHOULD fail on Mono if the runtime is doing something wrong - that's what the tests are for! Ignoring the tests implies that everything is fine, when it's not.